+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+- Minor: Added standalone backup API `pajlada::Settings::Backup::saveWithBackup` in `backup.hpp` (#90)
+
 ## v0.2.0
 
 - Breaking: Remove move ctor/operator for SettingListener. (#32)